configure.ac: Introduce the notion of android
This change is targeted towards detecting whether connman is compiled
for android. It does this by finding if host option passed to configure
has "android" in it. If it finds it makes two variables ANDROID and
ANDROID_COMPILE as 1. The first variable can be used in Makefile.am for
including different set of libraries (because of difference between
bionic and glibc). The second variable, ANDROID_COMPILE gets added into
config.h and then can be used in C code to conditionally compile android
specific code.