resolver: Adding ns_parse.c for android compilation
gweb/gresolv.c calls ns_initparse and few other ns_* functions. These
functions are part of resolver library in glibc. Bionic libc.so includes
libc_dns.a as a static library. These functions are defined as a part of
libc_dns which is compiled as static library but this library is not
installed as a part of android build. This commit adds a file called
ns_parse.c in gweb directory and implements whatever connman requires.
Makefile.am is modified to include this file only if it detects that
connman is built in Android environment.