gweb/gresolv.c: Adding android specific code in gresolv.c
Some of the structures are defined differently between glibc and bionic.
For example struct __res_state is defined in resolv_private.h in bionic
instead of resolv.h in glibc. __res_state definition is also different
between bionic and glibc. In glibc, this structure directly contains
nsaddrs which is an array of pointers of type sockaddr_in6. In bionic
ipv6 addresses are handled through an extension structure called struct
__res_state_ext. This commit addresses this difference by approopriately
accessing the members by using a compile time flag ANDROID_COMPILE which
is defined in config.h, created by configure process.